(#91): search Page 생성
