+++ src/features/tourList/ui/index.ts
@@ -1,4 +1,3 @@
-export { default as SearchForm } from './SearchForm';
 export { default as TourCardImages } from './TourCardImages';
 export { default as TourInfoCard } from './TourInfoCard';
 export { default as TouristFilterQueryUpdater } from './TouristFilterQueryUpdater';
