+++ src/features/tourSearch/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as SearchNavigate } from './SearchNavigate';
+export { default as SearchForm } from './SearchForm';
