+++ src/shared/ui/index.ts
@@ -9,3 +9,4 @@ export { default as Portal } from './Portal';
 export { default as TourCard } from './TourCard';
 export { default as TouristContentsTypeFilter } from './TouristContentsTypeFilter';
 export { default as TourTypeBadge } from './TourTypeBadge';
+export { default as BottomNavigationBar } from './bottomNavigationBar';
